Figured out how to do triangulation call.
The CSG library requiires all input faces to be convex polygons.
But the floor can have a concave shape.  So the thing to do is
to break the floor up into triangles, then I can use the CSG to union
all the triangular pieces into a clipping region for the individual
pieces of wood.